+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.2.43 (unreleased)
 -------------------
 
+- SUP-46571: Correct handling of "min" range for event recurrences
+  [boulch]
+
 - CITI-10 : Add (de) translation for event_public taxonomy and for event types vocabulary
   [boulch]